+++ src/maxpool_layer.h
@@ -6,14 +6,14 @@
 typedef struct {
     int h,w,c;
     int stride;
-    double *delta;
-    double *output;
+    float *delta;
+    float *output;
 } maxpool_layer;
 
 image get_maxpool_image(maxpool_layer layer);
 maxpool_layer *make_maxpool_layer(int h, int w, int c, int stride);
-void forward_maxpool_layer(const maxpool_layer layer, double *in);
-void backward_maxpool_layer(const maxpool_layer layer, double *in, double *delta);
+void forward_maxpool_layer(const maxpool_layer layer, float *in);
+void backward_maxpool_layer(const maxpool_layer layer, float *in, float *delta);
 
 #endif
