+++ src/softmax_layer.h
@@ -3,12 +3,12 @@
 
 typedef struct {
     int inputs;
-    double *delta;
-    double *output;
+    float *delta;
+    float *output;
 } softmax_layer;
 
 softmax_layer *make_softmax_layer(int inputs);
-void forward_softmax_layer(const softmax_layer layer, double *input);
-void backward_softmax_layer(const softmax_layer layer, double *input, double *delta);
+void forward_softmax_layer(const softmax_layer layer, float *input);
+void backward_softmax_layer(const softmax_layer layer, float *input, float *delta);
 
 #endif
